+++ example_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"fmt"
 
-	"github.com/vmihailenco/msgpack/v5"
+	"github.com/myhyh/msgpack/v5"
 )
 
 func ExampleMarshal() {