+++ extra/msgpappengine/appengine.go
@@ -3,7 +3,7 @@ package msgpappengine
 import (
 	"reflect"
 
-	"github.com/vmihailenco/msgpack/v5"
+	"github.com/myhyh/msgpack/v5"
 	ds "google.golang.org/appengine/datastore"
 )