+++ CHANGELOG.md
@@ -36,7 +36,7 @@
 
 ## v3
 
-- gopkg.in is not supported any more. Update import path to github.com/vmihailenco/msgpack.
+- gopkg.in is not supported any more. Update import path to github.com/myhyh/msgpack.
 - Msgpack maps are decoded into map[string]interface{} by default.
 - EncodeSliceLen is removed in favor of EncodeArrayLen. DecodeSliceLen is removed in favor of
   DecodeArrayLen.